TST Remove backend-specific 2D tests
This will make the 2D tests run for the new frontend, since the Scattering2D
interface hasn't changed. We keep these for now just to make sure we're
backwards compatible (since the default frontend is still the PyTorch one), but
once we're comfortable that the new, frontend-specific tests cover this, we
should remove these since they are redundant.